TrustLedger_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
Pillow
python-decouple
django-filter
